refactor(proxy): address review round (split teardown, dev-fallback method, addVolumes, fold project-url patching into shop config)
